Add integration test for changed and unchanged imports with an embedded nginx serving archives

+++ importer/src/main/java/life/catalogue/importer/ImportJob.java
@@ -211,14 +211,17 @@ public class ImportJob implements Runnable {
    * @return true if sourceDir should be imported
    */
   private boolean prepareSourceData(Path sourceDir) throws IOException, IllegalArgumentException, InterruptedException {
-    File archive = nCfg.archive(datasetKey, getAttempt());
+    final File archive = nCfg.archive(datasetKey, getAttempt());
+    final Path archivePath = archive.toPath();
+    final File lastArchive = dataset.getImportAttempt() == null ? null : nCfg.archive(datasetKey, dataset.getImportAttempt());
     archive.getParentFile().mkdirs();
 
+    boolean isModified = true;
     if (req.reimportAttempt != null) {
       // copy previous up/downloaded archive to repository
       Path prev = nCfg.archive(datasetKey, req.reimportAttempt).toPath();
       LOG.info("Symlink previous archive from import attempt {} for dataset {}: {}", req.reimportAttempt, datasetKey, prev);
-      Files.createSymbolicLink(archive.toPath(), prev);
+      Files.createSymbolicLink(archivePath, prev);
 
     } else if (req.hasUpload()) {
       // if data was uploaded we need to find out the format.
@@ -230,7 +233,7 @@ public class ImportJob implements Runnable {
       } else {
         // copy uploaded data to repository
         LOG.info("Move upload for dataset {} from {} to {}", datasetKey, req.upload, archive);
-        Files.move(req.upload, archive.toPath(), StandardCopyOption.REPLACE_EXISTING);
+        Files.move(req.upload, archivePath, StandardCopyOption.REPLACE_EXISTING);
       }
 
     } else if (DatasetOrigin.EXTERNAL == dataset.getOrigin()){
@@ -241,19 +244,26 @@ public class ImportJob implements Runnable {
         setFormat(proxy.format);
       } else {
         // download archive directly
+        boolean downloaded;
         if (req.force) {
           LOG.info("Force download of source for dataset {} from {} to {}", datasetKey, dataset.getDataAccess(), archive);
           downloader.download(di.getDownloadUri(), archive);
+          downloaded = true;
         } else {
           LOG.info("Download source for dataset {} from {} to {}", datasetKey, dataset.getDataAccess(), archive);
-          downloader.downloadIfModified(di.getDownloadUri(), archive);
+          var lmod = DownloadUtil.lastModified(lastArchive);
+          downloaded = downloader.downloadIfModified(di.getDownloadUri(), archive, lmod);
         }
         // make sure we received a real archive and not just a plain text error response
-        // https://github.com/CatalogueOfLife/backend/issues/1327
-        var size = Files.size(archive.toPath());
-        // 22 bytes is the minimum zip file length, BDJ responds with 14 text bytes only
-        if (size < 22) {
-          throw new DownloadException(di.getDownloadUri(), "Tiny file with "+size+" bytes. Cannot be an archive");
+        if (downloaded) {
+          verifyDownload(archivePath);
+        } else if (lastArchive != null){
+          isModified = false;
+          // symlink last one in case we want to force imports
+          Files.createSymbolicLink(archivePath, lastArchive.toPath());
+        } else {
+          // no download and no prev file should never happen!
+          throw new IllegalStateException("Download for dataset " + datasetKey + " does not exist. Skip import");
         }
       }
 
@@ -262,20 +272,19 @@ public class ImportJob implements Runnable {
       throw new IllegalStateException("Dataset " + datasetKey + " is not external and there are no uploads to be imported");
     }
 
-
+    // in case we haven't downloaded anything and this method returns false for not modified,
+    // the parent ImportJob will deleted this import attempt alltogether, so wrong MD5s from symlinks don't matter
     di.setMd5(ChecksumUtils.getMD5Checksum(archive));
-    di.setDownload(downloader.lastModified(archive));
+    di.setDownload(DownloadUtil.lastModified(archive));
     dao.update(di);
 
-    boolean isModified = true;
-    final Path archivePath = archive.toPath();
-    if (dataset.getImportAttempt() != null) {
+    // if we have a new archive test its MD5 to see if anything has changed
+    if (isModified && dataset.getImportAttempt() != null) {
       try (SqlSession session = factory.openSession()) {
-        String lastMD5 = session.getMapper(DatasetImportMapper.class).getMD5(datasetKey, dataset.getImportAttempt());
+        final String lastMD5 = session.getMapper(DatasetImportMapper.class).getMD5(datasetKey, dataset.getImportAttempt());
         if (Objects.equals(lastMD5, di.getMd5())) {
           // replace archive with symlink to last archive to save space
-          File lastArchive = nCfg.archive(datasetKey, dataset.getImportAttempt());
-          if (lastArchive.exists()) {
+          if (lastArchive != null && lastArchive.exists()) {
             // we have seen wrong MD5 hashes being stored (BDJ), so lets recalculate the last one from the file to make sure!
             var lastMD5Redone = ChecksumUtils.getMD5Checksum(lastArchive);
             if (lastMD5Redone.equals(lastMD5)) {
@@ -303,19 +312,6 @@ public class ImportJob implements Runnable {
       }
     }
 
-    // update latest symlink
-    Path latest = nCfg.lastestArchiveSymlink(datasetKey).toPath();
-    try {
-      Files.delete(latest);
-    } catch (NoSuchFileException e) {
-      // ignore, we dont want it anyways
-    }
-    try {
-      Files.createSymbolicLink(latest, archivePath);
-    } catch (IOException e) {
-      LOG.error("Failed to create latest symlink {} to archive {}: {}", latest, archivePath, e.getMessage(), e);
-    }
-
     checkIfCancelled();
     // decompress and import?
     if (isModified || req.force) {
@@ -330,9 +326,50 @@ public class ImportJob implements Runnable {
         setFormat(DataFormatDetector.detectFormat(sourceDir));
         LOG.info("Detected data format {} for dataset {}", dataset.getDataFormat(), dataset.getKey());
       }
+      // update latest symlink
+      updateLatestSymlink(archivePath);
       return true;
+
+    } else {
+      // we don't want any import and gonna delete the attempt completely.
+      LOG.info("Dataset {} sources unchanged. Stop import", datasetKey);
+      dao.delete(datasetKey, di.getAttempt());
+      if (archive.exists()) {
+        FileUtils.deleteQuietly(archive);
+      }
+      return false;
     }
-    return false;
+  }
+
+  private void updateLatestSymlink(Path archivePath) throws IOException {
+    Path latest = nCfg.lastestArchiveSymlink(datasetKey).toPath();
+    try {
+      Files.delete(latest);
+    } catch (NoSuchFileException e) {
+      // ignore, we dont want it anyways
+    }
+    try {
+      Files.createSymbolicLink(latest, archivePath);
+    } catch (IOException e) {
+      LOG.error("Failed to create latest symlink {} to archive {}: {}", latest, archivePath, e.getMessage(), e);
+    }
+  }
+
+  private void verifyDownload(Path archive) throws DownloadException {
+    if (!Files.exists(archive)) {
+      throw new DownloadException(di.getDownloadUri(), "Download file is missing: " + archive);
+    }
+    // https://github.com/CatalogueOfLife/backend/issues/1327
+    long size;
+    try {
+      size = Files.size(archive);
+    } catch (IOException e) {
+      throw new DownloadException(di.getDownloadUri(), e);
+    }
+    // 22 bytes is the minimum zip file length, BDJ responds with 14 text bytes only
+    if (size < 22) {
+      throw new DownloadException(di.getDownloadUri(), "Tiny file with "+size+" bytes. Cannot be a real archive");
+    }
   }
 
   private void importDataset() throws Exception {
@@ -391,10 +428,6 @@ public class ImportJob implements Runnable {
           di.setError(null);
           updateState(ImportState.FINISHED);
         }
-
-      } else {
-        LOG.info("Dataset {} sources unchanged. Stop import", datasetKey);
-        dao.delete(datasetKey, di.getAttempt());
       }
   
       if (iCfg.wait > 0) {
@@ -424,7 +457,9 @@ public class ImportJob implements Runnable {
       final File scratchDir = nCfg.scratchDir(datasetKey);
       LOG.debug("Remove scratch dir {}", scratchDir.getAbsolutePath());
       try {
-        PathUtils.deleteDirectory(sourceDir); // this is nested in the one below, but we see stale open files piling up
+        if (Files.exists(sourceDir)) {
+          PathUtils.deleteDirectory(sourceDir); // this is nested in the one below, but we see stale open files piling up
+        }
         FileUtils.deleteDirectory(scratchDir);
       } catch (IOException e) {
         LOG.error("Failed to remove scratch dir {}", scratchDir, e);
